Balance & Misc Changes
Added GridPBR shader (pretty much same as grid except allows shadows)
Battery now magnetizes within a longer range
Adjusted gun and player primitives
Enemy will now drop a variable number of batteries
Reorganized script organization
Made enemies red, player green

+++ Assets/Scripts/Managers/Spawner.cs
@@ -6,7 +6,7 @@ public class Spawner : MonoBehaviour
 
     private Transform player = null;
     private float nextSpawnTimer = 0f;
-    private float nextSpawnID = 0;
+    private int nextSpawnID = 0;
 
     private void Start()
     {
